requirements.txt
moviepy
playsound
pillow
numpy
